requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pdfplumber==0.11.4
python-docx==1.1.2
gunicorn==22.0.0
